Add files via upload
A weather detection website is built using HTML, CSS and JAVA SCRIPT where someone can mention the name of any city around the world and can get to see the current weather, humidity and precipitation chances of that city. The weather app fetches data from an API which is collected from the OpenWeather website.